requirements.txt
black
flask
werkzeug
django
psycopg2-binary
pytest
djangorestframework
djangorestframework-jsonapi
django-prometheus
prometheus_client
psutil
xenon
